pagination done

+++ src/State/actioncreators/posts.js
@@ -1,20 +1,37 @@
 import * as api from "../../api/index"
-import { CREATE, DELETE, FETCH_ALL, LIKE, UPDATE } from "../../constents/actionconstent";
-export let getPost=()=>async (dispatch)=>{
+import { CREATE, DELETE, FETCH_ALL, FETCH_BY_SEARCH, LIKE, UPDATE } from "../../constents/actionconstent";
+export let getPost=(page)=>async (dispatch)=>{
     try {
-        let {data}= await api.fetchposts();
-        dispatch({type:FETCH_ALL,payload:data.data});
+        let {data}= await api.fetchposts(page);
+        // console.log(data);
+        dispatch({type:FETCH_ALL,payload:data});
     } catch (error) {
         console.log("error");
     }
     // dispatch(action)
 }
 
+export let getPostBySearch=({searchquery,showalert})=>async (dispatch)=>{
+    try {
+        console.log(searchquery);
+        let {data}=await api.featchpostbysearch(searchquery);
+        // console.log("search",data.data);
+        if(!data.success){
+            showalert("error",data.data)
+        }
+        else{
+        dispatch({type:FETCH_BY_SEARCH,payload:data})
+        }
+    } catch (error) {
+        console.log(error);
+    }
+}
+
 export let creatPost=(Post)=>async (dispatch)=>{
     try{
-        console.log("post",Post);
+        // console.log("post",Post);
         let {data}=await api.creatPost(Post);
-        console.log("here",data)
+        // console.log("here",data)
         dispatch({type:CREATE,payload:data.data});
         dispatch(getPost());   
     }
@@ -32,11 +49,18 @@ export let updatepost=(id,newpost)=> async (dispatch)=>{
     }
 }
 
-export let deletepost=(id)=>async (dispatch)=>{
+export let deletepost=(id,showalert)=>async (dispatch)=>{
     try {
         let {data}=await api.deletepost(id);
         console.log(data);
-        dispatch({type:DELETE,payload:id})
+        if(data.success){
+            // alert(data.data)
+            showalert("error",data.data)
+            dispatch({type:DELETE,payload:id})
+        }
+        else{
+            showalert("error","could not delete post")
+        }
     } catch (error) {
         console.log(error);
     }
